# api/todo.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "todo.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/todos": {
      "get": {
        "operationId": "ListTodos",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/todoapp_grapiListTodosResponse"
            }
          }
        },
        "tags": [
          "TodoService"
        ]
      },
      "post": {
        "operationId": "CreateTodo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/todoapp_grapiTodo"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/todoapp_grapiTodo"
            }
          }
        ],
        "tags": [
          "TodoService"
        ]
      }
    },
    "/todos/{todo.todo_id}": {
      "patch": {
        "operationId": "UpdateTodo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/todoapp_grapiTodo"
            }
          }
        },
        "parameters": [
          {
            "name": "todo.todo_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/todoapp_grapiTodo"
            }
          }
        ],
        "tags": [
          "TodoService"
        ]
      }
    },
    "/todos/{todo_id}": {
      "get": {
        "operationId": "GetTodo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/todoapp_grapiTodo"
            }
          }
        },
        "parameters": [
          {
            "name": "todo_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "TodoService"
        ]
      },
      "delete": {
        "operationId": "DeleteTodo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          }
        },
        "parameters": [
          {
            "name": "todo_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "TodoService"
        ]
      }
    }
  },
  "definitions": {
    "todoapp_grapiListTodosResponse": {
      "type": "object",
      "properties": {
        "todos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/todoapp_grapiTodo"
          }
        }
      }
    },
    "todoapp_grapiTodo": {
      "type": "object",
      "properties": {
        "todo_id": {
          "type": "string"
        }
      }
    }
  }
}